+++ trunk/prj.sandbox/testgen/src/points2points_2d.cpp
@@ -47,13 +47,13 @@ void  testgen_points2points_2d( string res_folder )
     ofstream out((test_name + ".txt").c_str());
 
     out << num_clusters << endl;
-    for (int i=0; i<clusters.size(); i++)
+    for (int i=0; i<(int)clusters.size(); i++)
     {
       circle( res, clusters[i].first, clusters[i].second, Scalar(128,0,0,0), 2 );
       out <<  clusters[i].first.x << "\t" <<  clusters[i].first.y  << "\t" <<  clusters[i].second << endl;
     }
     out << p.size() << endl; // количество точек
-    for (int i = 0; i < p.size(); ++i)
+    for (int i = 0; i < (int)p.size(); ++i)
     {
       out << p[i].x << " " << p[i].y << endl;
       //circle( res, p[i], 2, Scalar(0, 0, 0), 2); //сами точки
